+++ tests/TestCase/Model/Table/AppTableTest.php
@@ -119,7 +119,7 @@ class AppTableTest extends TableTestCase
             $this->assertTrue($entity->get('active') && !$entity->get('created')->isFuture());
         }
 
-        $this->assertSqlEndsWith('FROM posts Posts WHERE (Posts.active = :c0 AND Posts.created <= :c1)', $query->sql());
+        $this->assertStringEndsWith('FROM posts Posts WHERE (Posts.active = :c0 AND Posts.created <= :c1)', $query->sql());
         $this->assertTrue($query->getValueBinder()->bindings()[':c0']['value']);
         $this->assertInstanceOf(FrozenTime::class, $query->getValueBinder()->bindings()[':c1']['value']);
     }
@@ -136,7 +136,7 @@ class AppTableTest extends TableTestCase
             $this->assertTrue(!$entity->get('active') || $entity->get('created')->isFuture());
         }
 
-        $this->assertSqlEndsWith('FROM posts Posts WHERE (Posts.active = :c0 OR Posts.created > :c1)', $query->sql());
+        $this->assertStringEndsWith('FROM posts Posts WHERE (Posts.active = :c0 OR Posts.created > :c1)', $query->sql());
         $this->assertFalse($query->getValueBinder()->bindings()[':c0']['value']);
         $this->assertInstanceOf(FrozenTime::class, $query->getValueBinder()->bindings()[':c1']['value']);
     }
@@ -148,10 +148,10 @@ class AppTableTest extends TableTestCase
     public function testFindRandomMethod(): void
     {
         $query = $this->Posts->find('random');
-        $this->assertSqlEndsWith('FROM posts Posts ORDER BY rand() LIMIT 1', $query->sql());
+        $this->assertStringEndsWith('FROM posts Posts ORDER BY rand() LIMIT 1', $query->sql());
 
         $query = $this->Posts->find('random')->limit(2);
-        $this->assertSqlEndsWith('FROM posts Posts ORDER BY rand() LIMIT 2', $query->sql());
+        $this->assertStringEndsWith('FROM posts Posts ORDER BY rand() LIMIT 2', $query->sql());
     }
 
     /**
@@ -185,7 +185,7 @@ class AppTableTest extends TableTestCase
         $this->assertNotEmpty($query->toArray());
         $fromCache = Cache::read('posts_list', $this->Posts->getCacheName())->toArray();
         $this->assertEquals($query->toArray(), $fromCache);
-        $this->assertSqlEndsWith('ORDER BY ' . $this->Posts->getDisplayField() . ' ASC', $query->sql());
+        $this->assertStringEndsWith('ORDER BY ' . $this->Posts->getDisplayField() . ' ASC', $query->sql());
     }
 
     /**
@@ -201,7 +201,7 @@ class AppTableTest extends TableTestCase
             2 => 'Another post category',
         ];
         $query = $this->Posts->Categories->getTreeList();
-        $this->assertSqlEndsNotWith('ORDER BY `' . $this->Posts->Categories->getDisplayField() . '` ASC', $query->sql());
+        $this->assertStringEndsNotWith('ORDER BY `' . $this->Posts->Categories->getDisplayField() . '` ASC', $query->sql());
         $this->assertEquals($expected, $query->toArray());
         $fromCache = Cache::read('posts_categories_tree_list', $this->Posts->Categories->getCacheName())->toArray();
         $this->assertEquals($query->toArray(), $fromCache);
@@ -249,13 +249,13 @@ class AppTableTest extends TableTestCase
             'created' => '2016-12',
         ];
         $query = $this->Posts->queryFromFilter($this->Posts->find(), $data);
-        $this->assertSqlEndsWith($expectedSql, $query->sql());
+        $this->assertStringEndsWith($expectedSql, $query->sql());
 
         $params = array_map(fn($value) => $value instanceof I18nDateTimeInterface ? $value->i18nFormat() : $value, collection($query->getValueBinder()->bindings())->extract('value')->toList());
         $this->assertEquals($expectedParams, $params);
 
         $query = $this->Posts->queryFromFilter($this->Posts->find(), ['active' => I18N_NO] + $data);
-        $this->assertSqlEndsWith($expectedSql, $query->sql());
+        $this->assertStringEndsWith($expectedSql, $query->sql());
         $this->assertFalse($query->getValueBinder()->bindings()[':c5']['value']);
 
         //With some invalid data