+++ tests/TestCase/Model/Table/UsersTableTest.php
@@ -152,7 +152,7 @@ class UsersTableTest extends TableTestCase
     public function testFindActive(): void
     {
         $query = $this->Table->find('active');
-        $this->assertSqlEndsWith('FROM users Users WHERE (active = :c0 AND banned = :c1)', $query->sql());
+        $this->assertStringEndsWith('FROM users Users WHERE (active = :c0 AND banned = :c1)', $query->sql());
         $this->assertTrue($query->getValueBinder()->bindings()[':c0']['value']);
         $this->assertFalse($query->getValueBinder()->bindings()[':c1']['value']);
     }
@@ -164,7 +164,7 @@ class UsersTableTest extends TableTestCase
     public function testFindAuth(): void
     {
         $query = $this->Table->find('auth');
-        $this->assertSqlEndsWith('FROM `users` `Users` INNER JOIN `users_groups` `Groups` ON `Groups`.`id` = `Users`.`group_id`', $query->sql());
+        $this->assertStringEndsWith('FROM users Users INNER JOIN users_groups Groups ON Groups.id = Users.group_id', $query->sql());
     }
 
     /**
@@ -174,7 +174,7 @@ class UsersTableTest extends TableTestCase
     public function testFindBanned(): void
     {
         $query = $this->Table->find('banned');
-        $this->assertSqlEndsWith('FROM users Users WHERE banned = :c0', $query->sql());
+        $this->assertStringEndsWith('FROM users Users WHERE banned = :c0', $query->sql());
         $this->assertTrue($query->getValueBinder()->bindings()[':c0']['value']);
     }
 
@@ -185,7 +185,7 @@ class UsersTableTest extends TableTestCase
     public function testFindPending(): void
     {
         $query = $this->Table->find('pending');
-        $this->assertSqlEndsWith('FROM users Users WHERE (active = :c0 AND banned = :c1)', $query->sql());
+        $this->assertStringEndsWith('FROM users Users WHERE (active = :c0 AND banned = :c1)', $query->sql());
         $this->assertFalse($query->getValueBinder()->bindings()[':c0']['value']);
         $this->assertFalse($query->getValueBinder()->bindings()[':c1']['value']);
     }
@@ -197,7 +197,7 @@ class UsersTableTest extends TableTestCase
     public function testGetActiveList(): void
     {
         $query = $this->Table->getActiveList();
-        $this->assertSqlEndsWith('FROM users Users WHERE active = :c0 ORDER BY username ASC', $query->sql());
+        $this->assertStringEndsWith('FROM users Users WHERE active = :c0 ORDER BY username ASC', $query->sql());
         $this->assertNotEmpty($query->toArray());
         $fromCache = Cache::read('active_users_list', $this->Table->getCacheName())->toArray();
         $this->assertEquals($fromCache, $query->toArray());
@@ -216,21 +216,21 @@ class UsersTableTest extends TableTestCase
         ];
 
         $query = $this->Table->queryFromFilter($this->Table->find(), $data);
-        $this->assertSqlEndsWith('FROM users Users WHERE (username like :c0 AND group_id = :c1 AND active = :c2 AND banned = :c3)', $query->sql());
+        $this->assertStringEndsWith('FROM users Users WHERE (username like :c0 AND group_id = :c1 AND active = :c2 AND banned = :c3)', $query->sql());
         $this->assertEquals('%test%', $query->getValueBinder()->bindings()[':c0']['value']);
         $this->assertEquals(1, $query->getValueBinder()->bindings()[':c1']['value']);
         $this->assertTrue($query->getValueBinder()->bindings()[':c2']['value']);
         $this->assertFalse($query->getValueBinder()->bindings()[':c3']['value']);
 
         $query = $this->Table->queryFromFilter($this->Table->find(), ['status' => 'pending'] + $data);
-        $this->assertSqlEndsWith('FROM users Users WHERE (username like :c0 AND group_id = :c1 AND active = :c2 AND banned = :c3)', $query->sql());
+        $this->assertStringEndsWith('FROM users Users WHERE (username like :c0 AND group_id = :c1 AND active = :c2 AND banned = :c3)', $query->sql());
         $this->assertEquals('%test%', $query->getValueBinder()->bindings()[':c0']['value']);
         $this->assertEquals(1, $query->getValueBinder()->bindings()[':c1']['value']);
         $this->assertFalse($query->getValueBinder()->bindings()[':c2']['value']);
         $this->assertFalse($query->getValueBinder()->bindings()[':c3']['value']);
 
         $query = $this->Table->queryFromFilter($this->Table->find(), ['status' => 'banned'] + $data);
-        $this->assertSqlEndsWith('FROM users Users WHERE (username like :c0 AND group_id = :c1 AND banned = :c2)', $query->sql());
+        $this->assertStringEndsWith('FROM users Users WHERE (username like :c0 AND group_id = :c1 AND banned = :c2)', $query->sql());
         $this->assertEquals('%test%', $query->getValueBinder()->bindings()[':c0']['value']);
         $this->assertEquals(1, $query->getValueBinder()->bindings()[':c1']['value']);
         $this->assertTrue($query->getValueBinder()->bindings()[':c2']['value']);