+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### 2.32.1
 * improved the query filter for posts and pages in the admin panel: now the `title` field works for both title and slug;
 * `User::_getLastLogins()` returns an array and no longer a collection;
+* `assertSqlEndsWith()` and `assertSqlEndsNotWith()` methods have been replaced with `assertStringEndsWith()` and 
+  `assertStringEndsNotWith()`. The `quoteIdentifiers` option for tests has been disabled;
 * updated for php-tools 1.7.4.
 
 ### 2.32.0